# wavefront/serializer: improve performance by ~30%