@danielnelson 2019-05-15
@puckpuck Can you review?  I assume any changes would also need to go into the wavefront-sdk.